Fix building on kernels 6.4.0 and above
On these kernels, kernel_termios_to_user_termios is a proper function which
is not exported, so we can't use it. We assume that structs termios2 and
ktermios2 are bit-identical and use copy_to_user instead.

This assumption isn't correct on sparc, so we trigger a build error there.

Based on a commit by Dmytro Bagrii.